# openapi.json
{
    "openapi": "3.1.0",
    "info": {
      "title": "Premier League Clubs",
      "version": "1.0.0"
    },
    "paths": {
...
        "/igraci": {
            "get": {
                "summary": "Get players",
                "description": "Returns all players in collection",
                "produces": [
                "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Fetch successful"
                    },
                    "500":{
                        "description": "Server error"
                    }
                }
            }
        },
...
        "/{imekluba}/wiki": {
            "get":{
                "summary": "Get specified club's wiki handle",
                "description": "Returns specified club's wiki handle from collection",
                "produces": [
                "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Fetch successful"
                    },
                    "404":{
                        "description": "Bad request"
                    }
                }
            }
        },
...
    }
}